Botões play/stop e contador funcionando individualmente, download de audio baixado na pasta de downloads
