+++ lib/Screens/AudioTextDiaryPage/TextDiary/text_diary.dart
